+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 
 ## Unreleased
 
+## [24.4.0] - 2020-08-31
+
 ## [24.3.0] - 2020-08-25
 
 ## [24.2.3] - 2020-08-19